wish me luck

+++ src/org/psnbtech/Clock.java
@@ -2,39 +2,39 @@ package org.psnbtech;
 
 /**
  * The {@code Clock} class is responsible for tracking the number of cycles
- * that have elapsed over time. 
+ * that have elapsed over time.
  * @author Brendan Jones
  *
  */
 
 public class Clock {
-	
+
 	/**
 	 * The number of milliseconds that make up one cycle.
 	 */
 	private float millisPerCycle;
-	
+
 	/**
 	 * The last time that the clock was updated (used for calculating the
 	 * delta time).
 	 */
 	private long lastUpdate;
-	
+
 	/**
 	 * The number of cycles that have elapsed and have not yet been polled.
 	 */
 	private int elapsedCycles;
-	
+
 	/**
 	 * The amount of excess time towards the next elapsed cycle.
 	 */
 	private float excessCycles;
-	
+
 	/**
 	 * Whether or not the clock is paused.
 	 */
 	private boolean isPaused;
-	
+
 	/**
 	 * Creates a new clock and sets it's cycles-per-second.
 	 * @param cyclesPerSecond The number of cycles that elapse per second.
@@ -43,7 +43,7 @@ public class Clock {
 		setCyclesPerSecond(cyclesPerSecond);
 		reset();
 	}
-	
+
 	/**
 	 * Sets the number of cycles that elapse per second.
 	 * @param cyclesPerSecond The number of cycles per second.
@@ -51,7 +51,7 @@ public class Clock {
 	public void setCyclesPerSecond(float cyclesPerSecond) {
 		this.millisPerCycle = (1.0f / cyclesPerSecond) * 1000;
 	}
-	
+
 	/**
 	 * Resets the clock stats. Elapsed cycles and cycle excess will be reset
 	 * to 0, the last update time will be reset to the current time, and the
@@ -63,7 +63,7 @@ public class Clock {
 		this.lastUpdate = getCurrentTime();
 		this.isPaused = false;
 	}
-	
+
 	/**
 	 * Updates the clock stats. The number of elapsed cycles, as well as the
 	 * cycle excess will be calculated only if the clock is not paused. This
@@ -73,18 +73,18 @@ public class Clock {
 	public void update() {
 		//Get the current time and calculate the delta time.
 		long currUpdate = getCurrentTime();
-		float delta = (float)(currUpdate - lastUpdate) + excessCycles;
-		
+		float delta = currUpdate - lastUpdate + excessCycles;
+
 		//Update the number of elapsed and excess ticks if we're not paused.
 		if(!isPaused) {
 			this.elapsedCycles += (int)Math.floor(delta / millisPerCycle);
 			this.excessCycles = delta % millisPerCycle;
 		}
-		
+
 		//Set the last update time for the next update cycle.
 		this.lastUpdate = currUpdate;
 	}
-	
+
 	/**
 	 * Pauses or unpauses the clock. While paused, a clock will not update
 	 * elapsed cycles or cycle excess, though the {@code update} method should
@@ -94,7 +94,7 @@ public class Clock {
 	public void setPaused(boolean paused) {
 		this.isPaused = paused;
 	}
-	
+
 	/**
 	 * Checks to see if the clock is currently paused.
 	 * @return Whether or not this clock is paused.
@@ -102,7 +102,7 @@ public class Clock {
 	public boolean isPaused() {
 		return isPaused;
 	}
-	
+
 	/**
 	 * Checks to see if a cycle has elapsed for this clock yet. If so,
 	 * the number of elapsed cycles will be decremented by one.
@@ -116,7 +116,7 @@ public class Clock {
 		}
 		return false;
 	}
-	
+
 	/**
 	 * Checks to see if a cycle has elapsed for this clock yet. Unlike
 	 * {@code hasElapsedCycle}, the number of cycles will not be decremented
@@ -127,7 +127,7 @@ public class Clock {
 	public boolean peekElapsedCycle() {
 		return (elapsedCycles > 0);
 	}
-	
+
 	/**
 	 * Calculates the current time in milliseconds using the computer's high
 	 * resolution clock. This is much more reliable than
